refactor: moved p2p chat related hooks to its own entity
