+++ packages/api/src/hooks/index.ts
@@ -20,7 +20,7 @@ export { default as useCFDCompareAccounts } from './useCFDCompareAccounts';
 export { default as useCancelCryptoTransaction } from './useCancelCryptoTransaction';
 export { default as useCashierFiatAddress } from './useCashierFiatAddress';
 // TODO: Export these into P2P entities
-export { default as useChatCreate } from './p2p/useChatCreate';
+export { default as useChatCreate } from './p2p/entity/chat/p2p-chat/useChatCreate';
 export { default as useAdvertiserInfo } from './p2p/useAdvertiserInfo';
 export { default as useOrderInfo } from './p2p/useOrderInfo';
