+++ packages/api/src/hooks/p2p/entity/chat/index.ts
@@ -1,0 +1,1 @@
+export * as chat from './p2p-chat';
