+++ packages/api/src/hooks/p2p/entity/chat/p2p-chat/index.ts
@@ -1,0 +1,1 @@
+export { default as useCreate } from './useChatCreate';
